Introduced the channel extractor in the multicore version and changed a type from int to size_t
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@222 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/Input_node.h
@@ -87,8 +87,8 @@ public:
   void remove_from_active_list(int stream);
 
   // Callback functions:
-  void hook_added_data_reader(int reader);
-  void hook_added_data_writer(int writer);
+  void hook_added_data_reader(size_t reader);
+  void hook_added_data_writer(size_t writer);
 
 private:
   // Stop a stream in active_list
